I added a distributed matrix class to libmints.
The new class distributes a 2D matrix.  The matrix is split
up into nproc blocks and each block is further subdivided into
sub_blocks. A few of the  current capabilities include inserting a value,
adding two matrices, and matrix multiply. This is still very
experimental. I am only adding it so I can put it on jaguar to
run a few tests. It only works with MADNESS, but it should
compile without MADNESS.  The constructors throw an exception
if PSI4 is not built with MADNESS.

git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@2158 6a214465-e6f0-41bc-91e6-de896bb96b5e

Former-commit-id: 51d64cf216598738f1939981b7ce174b0f5073ed