I added documentation to libdist_matrix.
The distributed matrix class is included in libdist_matrix. There
is user and programmer documentation that I have built using doxygen.
I will need to test wether it is compatible with the PSI4 documentation,
but I think that it will be.  I also changed some shared_ptr in src/bin/omp2
to boost::shared_ptr.