Fix the build with some std libraries. See comment.
It is a bad idea to use a name that starts with an underscore and a
lowercase letter for type names or defines as done in physconst.h. The
symbols starting with an underscore and lowercase letters are reserved
(see standard). So the define _c conflicts with my std library. At
least fix this by reordering includes so that physconst.h goes after
standard includes.