This commit allows to compile psi4 using clang++ and the new C++11 standard. I had to chop a few heads to make this work. 1) Libyeti is no longer compiled. 2) Configure does no longer check for isinf() or finite(). These two    functions are required by madness. 3) The const_iterator over vector was eliminated, and the non-const version    reworked because it was not compiling.
All these changes are not sufficient to compile psi4 directly with clang++.
Boost needs to be modified and compiled separately.

This bit of code must be added to boost/config/stdlib/libcpp.hpp
to make boost work with clang++

 #if _LIBCPP_VERSION < 1002
   define BOOST_NO_CXX11_NUMERIC_LIMITS
 #endif

In principle this could be added to our local copy of boost but
I decided it was better to not mess up with it.

I know, this was supposed to be a short message but it ended up
becoming a short essay.