Fixes compilation of LibERD as standalone library (i.e. make erd works)
Problem was in the -ffree-form (GNU) or -free (Intel) flags, as the code is not free-format Fortran.
The -fimplicit-none flag is OK (both ERD and the DKH code use it internally already)
Enforcement of the 2003 Fortran standard has been removed on the basis that the Fortran codes are not developed
by the Psi team but used as is.


Former-commit-id: 21c501d0504239cbe6a6e3663deaeb41b7625e2c