Fixed bug in CIS symmetry, fixed HDF5 detection.
The CISRHamiltonian.product, computing Hx = b was using the
wrong irrep for b in case x is a delta vector.

The Cmake HDF5 detection needed HDF5_ROOT to be an environment variable,
the newly included cmake/FindHDF5.cmake allows the --hdf5-dir option
from the setup to smoothly pass appropriate variables without
modifying the environment.
A typo corrected in cmake/ConfigCheMPS2.cmake that prevented
proper warning when HDF5 was not detected.


Former-commit-id: 9b17de0e3c5e5301e8db62c280029b46b14f9973