New UHF stability analysis code.
This code also can reoptimize the orbitals until the wavefunction
is stable.

Please note that this code relies on copy-pasting and is thus
very ugly. This is because I need it to work now, and also because
the solver and hamiltonian classes are not very flexible when it
comes to unrestricted orbitals.

The new classes are: UStabHamiltonian, to compute the product of
the electronic Hessian with an arbitrary vector for UHF, DLUSolver,
a Davidson solver that can work with unrestricted Hamiltonian,
and UStab, the stability analysis itself.

UHF now uses the new code for all cases, except when scf_type=pk
because pk integrals cannot handle non-symmetric pseudo-density
matrices. The old code is used for scf_type=pk, but in the end
we should probably just nuke it.


Former-commit-id: cb2e48b1614b16178148786cba212e6e94712a6e