Adds support for interfacing with Ambit.
Updates Psi4 to support interfacing with the Ambit tensor library. Also
provides a working plugin template.

This commit also renames the NotImplementedException macro to
NOT_IMPLEMENTED_EXCEPTION to avoid conflicts with Ambit.

*Note*: This commit has Ambit enabled by default. This was to have
Travis CI test the code. In a future commit this will be set to
disabled by default until such time it is determined to switch it back
to enabled by default.