Prototype OpenMP parallel code for PK integrals.
There is a lot of debug printing (commented out for now). The
PK integrals computation is now multithreaded, and the asynchronous
IO handler now supports being called from multiple threads.
Whenever a thread has a writing job, it is sent to the AIO handler
which uses a single thread to do all the writing.

The code needs clean-up and more optimization.


Former-commit-id: e55035de30508cf300532eecff0590fbb049142d