Fixed ROHF stability analysis and added test.
During the recent changes to the CC codes, we changed the ROHF behavior of libtrans, to push the socc orbitals to the end of the list of virtuals, rather than the start.  This patch uses the new ordering in the ROHF stabilty code, which had been broken for a few weeks.  Because there was no test, the problem went unnoticed; the stability2 test is designed to test ROHF stability.  The current algorithm should be replaced by a Davidson solver one fine day.


Former-commit-id: 396b100374bf1e3da3a4d12b99b12b2b46d4fa2b