Removes redundant code from roa.py adds back up of results
Replaces them with calls to generalized code in findif_response_util
module that will be used across all drivers for this type of code.

Adds check for roa_computed flag in db: useful for debugging. It will
copy the database, and make a backup of it. Changes to psi4.scatter()
will be printed in the outfile, and the old results retrievable from the
backup with a small amount of effort. Eventually I will make sure the
final results are stored in the database making it even easier to
retrieve them.