Make Psi4 buildable with Ninja
Ninja is supposed to be faster at compiling than make. CMake supports it
out-of-the-box, it's enough to pass it as generator:
  cmake -H. -Bbuild -GNinja

Modifications needed to get Ninja to work:
- DESTDIR=${CMAKE_BINARY_DIR}/stage was prepended to ${CMAKE_MAKE_PROGRAM} install appearing in all instances of INSTALL_COMMAND
- Set generator also for skeleton plugins in the add_skeleton_plugin
macro psi4OptionsTools.cmake

Caveats:
- The official Ninja distribution is not yet capable of handling
Fortran. It is possible to use the Kitware-maintained version of Ninja
for that though https://github.com/Kitware/ninja/releases