Rebuild denominators in CCENERGY
Running a macroiteration scheme like the one needed for PTED,
needs to have the denominators for CCENERGY updated at every
re-entry in the T amplitude equation solver.
This commit adds a rebuild_denom function that is called before
initializing the amplitudes, if a PCM-CC-PTED calculation is run.
Thanks @amjames and @ashutoshvt for help in spotting the problem.