Update dftd3.rst (#1085)
* Update dftd3.rst

Update links to Grimme's DFT-D3.  Note that the link to the parameter source may not be correct - those with knowledge of the source, please chime in.  Fixes 1084.

* Update dftd3.rst

Update links to Grimme's DFT-D3.  Note that the link to the parameter source may not be correct - those with knowledge of the source, please chime in.  Fixes 1084.

* Update gcp.rst

* doc: the sphinx wants blank lines