Remove third Travis test
The third Travis test is timing out too much.  This can be traced to the painfully slow download time for the GCC compiler.  We have plans to replace this test with some plugin validation, using native GCC compilers to avoid the timeouts.  For now, we'll just nuke it to allow the backlog of PRs through.