Ensure that initialize_lebedev is called
Previously, initialize_lebedev was never called, and in fact was getting
optimized out of the module completely upon compilation. When
lebedev_mappping_[] is then accessed across multiple OpenMP threads, the
std::map is empty, and a deadlock can happen where two threads try to
access-write (since [key] fills if key is not found), and the slightly
slower thread ends up in a Bad State where it thinks there is a value
but ends up infinitely looping on the lookup (the program will hang on
[]).

This only happens once every several thousand runs, and only when
running with a high degree of parallelism in a system with many atoms. I
cannot induce it in captivity, but I have observed it in the wild.

Anyway, [] accesses on std::map aren't thread-safe if you aren't
super-duper sure the map is fully filled for all keys you'd ever look
up, which *should* be the case if initialize_lebedev was ever called
anywhere. But it wasn't, and that was Bad.

Now it's called exactly once (thanks, c++11's call_once!).

The hangs should be gone, though I'll have to churn through another
several thousand runs to likely be sure (as, again, it is a very rare
kind of hang).

That said, as far as I can tell, besides one print function the
resulting order_ that's assigned to is never *used*. Maybe a candidate
to be axed in the future?